fix: overflowing add doesn't compile due to type error

+++ src/cpu/opcodes.rs
@@ -244,7 +244,7 @@ pub(super) fn set_sound_timer_to_vx(cpu: &mut CPU, x: usize) {
 pub(super) fn add_vx_to_index(cpu: &mut CPU, x: usize) {
     let i = cpu.index_register;
     let vx = cpu.reg[x];
-    let (sum, overflow) = i.overflowing_add(vx);
+    let (sum, overflow) = i.overflowing_add(vx as usize);
     let overflow = if overflow { 1 } else { 0 };
 
     cpu.index_register = sum;
